+++ src/Compute/Compute/VirtualMachine/Operation/NewAzureVMCommand.cs
@@ -258,7 +258,7 @@ namespace Microsoft.Azure.Commands.Compute
             HelpMessage = "EncryptionAtHost property can be used by user in the request to enable or disable the Host Encryption for the virtual machine. This will enable the encryption for all the disks including Resource/Temp disk at host itself.")]
         [Parameter(ParameterSetName = DiskFileParameterSet, Mandatory = false,
             HelpMessage = "EncryptionAtHost property can be used by user in the request to enable or disable the Host Encryption for the virtual machine. This will enable the encryption for all the disks including Resource/Temp disk at host itself.")]
-        public SwitchParameter EncryptionAtHost { get; set; } = false;
+        public SwitchParameter EncryptionAtHost { get; set; }
 
         public override void ExecuteCmdlet()
         {
@@ -395,7 +395,7 @@ namespace Microsoft.Azure.Commands.Compute
                         priority: _cmdlet.Priority,
                         evictionPolicy: _cmdlet.EvictionPolicy,
                         maxPrice: _cmdlet.IsParameterBound(c => c.MaxPrice) ? _cmdlet.MaxPrice : (double?)null,
-                        encryptionAtHostEnabled: _cmdlet.EncryptionAtHost.IsPresent
+                        encryptionAtHostPresent: _cmdlet.EncryptionAtHost.IsPresent
                         );
                 }
                 else
@@ -420,8 +420,8 @@ namespace Microsoft.Azure.Commands.Compute
                         priority: _cmdlet.Priority,
                         evictionPolicy: _cmdlet.EvictionPolicy,
                         maxPrice: _cmdlet.IsParameterBound(c => c.MaxPrice) ? _cmdlet.MaxPrice : (double?)null,
-                        encryptionAtHostEnabled: _cmdlet.EncryptionAtHost.IsPresent
-                        );
+                        encryptionAtHostPresent: _cmdlet.EncryptionAtHost.IsPresent
+                    );
                 }
             }
         }